backend/requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
bitarray==2.7.3
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
parsimonious==0.9.0
Pillow==9.5.0
psycopg2==2.9.6
pycryptodome==3.18.0
PyJWT==2.7.0
pytz==2023.3
redis==4.5.5
regex==2023.5.5
requests==2.31.0
rlp==3.0.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
sorl-thumbnail==12.9.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==2.0.2
